mainCamera/filter/legacy/web_demo/requirements.txt
# Fujifilm 軟片模擬系統相依套件

# 核心圖像處理
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0

# Web 應用程式
Flask>=2.0.0
Werkzeug>=2.0.0

# 可選套件 (用於進階功能)
matplotlib>=3.3.0  # 用於圖表和分析
scikit-image>=0.18.0  # 進階圖像處理

# Raspberry Pi 相關 (如果在 RPi 上運行)
# picamera2  # RPi 5 相機模組
# RPi.GPIO   # GPIO 控制

# 檔案格式支援
imageio>=2.9.0
